refactor: weather_screenとweather_imageをuiディレクトリへ移動
