+++ lib/ui/weather_image.dart
